+++ flutter/test/integrations/flutter_error_integration_test.dart
@@ -56,8 +56,6 @@ void main() {
       final throwableMechanism = event.throwableMechanism as ThrowableMechanism;
       expect(throwableMechanism.mechanism.type, 'FlutterError');
       expect(throwableMechanism.mechanism.handled, true);
-      expect(throwableMechanism.mechanism.data['hint'],
-          'See "flutter_error_details" down below for more information');
       expect(throwableMechanism.throwable, exception);
 
       expect(event.contexts['flutter_error_details']['library'], 'sentry');
@@ -92,8 +90,6 @@ void main() {
       final throwableMechanism = event.throwableMechanism as ThrowableMechanism;
       expect(throwableMechanism.mechanism.type, 'FlutterError');
       expect(throwableMechanism.mechanism.handled, true);
-      expect(throwableMechanism.mechanism.data['hint'],
-          'See "flutter_error_details" down below for more information');
 
       expect(event.contexts['flutter_error_details']['library'], 'sentry');
       expect(event.contexts['flutter_error_details']['context'],