+++ file/lib/sentry_file.dart
@@ -1,0 +1,1 @@
+export 'src/sentry_file.dart';